pull_client: test if files exists on delete step

+++ services/pulse2/pull_client/command.py
@@ -220,8 +220,10 @@ class DeleteStep(Step):
 
     def run(self):
         package_dir = os.path.join(get_packages_dir(), self.command.package_uuid)
-        shutil.rmtree(package_dir)
-        os.unlink(package_dir + ".zip")
+        if os.path.exists(package_dir):
+            shutil.rmtree(package_dir)
+        if os.path.exists(package_dir + ".zip"):
+            os.unlink(package_dir + ".zip")
         return ("%f O: Package files removed" % time.time(), 0)
